Clarified some logic. Made it 64bit happy.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@878 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/client/wad.c
@@ -384,7 +384,7 @@ qbyte *W_ConvertWAD3Texture(miptex_t *tex, int *width, int *height, qboolean *us
 	if (!data)
 		return NULL;
 
-	in = (qbyte *)((int) tex + tex->offsets[0]);
+	in = (qbyte *)tex + tex->offsets[0];
 
 	*width = tex->width;
 	*height = tex->height;
